Refactor fetcher methods into util class
